rename: 도메인별 패키지 이동
- Line, Station 으로 분리
