+++ src/main/java/subway/line/LineController.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.line;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
