+++ src/main/java/subway/line/LineRequest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.line;
 
 public class LineRequest {
     private String name;
