+++ src/main/java/subway/line/LineResponse.java
@@ -1,4 +1,6 @@
-package subway;
+package subway.line;
+
+import subway.station.StationResponse;
 
 import java.util.List;
