+++ src/main/java/subway/line/LineService.java
@@ -1,7 +1,10 @@
-package subway;
+package subway.line;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.station.Station;
+import subway.station.StationRepository;
+import subway.station.StationResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
