+++ src/main/java/subway/line/LineUpdateRequest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.line;
 
 public class LineUpdateRequest {
     private String name;
