+++ src/main/java/subway/station/StationLink.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import javax.persistence.Embeddable;
 import javax.persistence.FetchType;
